Add monitor for stopped services when clone=False
The cluster as its currently configured for services with clone=False,
Pacemaker will monitor exclusively that the daemon is running in the
node where it should, but will take no actions if the same daemon is
running (e.g. started manually by a sysadmin) in another node of the
cluster, this becomes a problem for services that are expected to be
configured in active/passive (e.g. manila-share).

This change configures two monitors for services with clone=False, one
that monitors the daemon is running where it should, and another one
that monitors the daemon is not running where it shouldn't.

primitive res_apache systemd:apache2 \
	...
        op monitor interval=5s role=Started \
        op monitor interval=6s role=Stopped

https://clusterlabs.org/pacemaker/doc/deprecated/en-US/Pacemaker/1.1/html/Pacemaker_Explained/_resource_operations.html#s-resource-monitoring

Closes-Bug: #1904623
Change-Id: I9e5383f5ab6b6967aa0f2318764519989a292227