Use the passed arguments in remove_dnsha
This appears to be a typo, where self.service_name and self.endpoint_type
was used in the same way as seen in other methods,
but self.service_name or self.endpoint_type is not defined in this class,
or any subclasses of it.
So we can use the passed name and endpoint_type arguments here.

Closes-Bug: #2037515
Change-Id: I12c561057372c11632b2d97fa1763fc92d89f479